+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-08-14 18:03:25
-// Commit: style(state): harmonize light and dark theme palette for state
+// Updated at: 2026-08-16 10:09:49
+// Commit: perf(state): optimize memory consumption in state list viewport
 
 class stateComponentState {
-  final String commitRef = '3616';
-  final String lastMessage = 'style(state): harmonize light and dark theme palette for state';
-  final DateTime timestamp = DateTime.parse('2026-08-14T18:03:25+05:00');
+  final String commitRef = '3624';
+  final String lastMessage = 'perf(state): optimize memory consumption in state list viewport';
+  final DateTime timestamp = DateTime.parse('2026-08-16T10:09:49+05:00');
 }